@@ -11,7 +11,7 @@ package example
 
 import (
 	"database/sql/driver"
-	"encoding/json"
+	json "encoding/json"
 	"errors"
 	"fmt"
 	"strings"
@@ -107,6 +107,14 @@ func (x *StrState) UnmarshalText(text []byte) error {
 	}
 	*x = tmp
 	return nil
+}
+
+// AppendText appends the textual representation of itself to the end of b
+// (allocating a larger slice if necessary) and returns the updated slice.
+//
+// Implementations must not retain b, nor mutate any bytes within b[:len(b)].
+func (x *StrState) AppendText(b []byte) ([]byte, error) {
+	return append(b, x.String()...), nil
 }
 
 var errStrStateNilPtr = errors.New("value pointer is nil") // one per type for package clashes
