@@ -13,7 +13,8 @@ import (
 )
 
 const (
-	testExample = `example_test.go`
+	testExample       = `example_test.go`
+	testExampleNoIota = `example_no_iota_test.go`
 )
 
 // TestNoStructInputFile
@@ -24,8 +25,7 @@ func TestNoStructFile(t *testing.T) {
 
 	}
 	`
-	g := NewGenerator().
-		WithoutSnakeToCamel()
+	g := NewGenerator(WithoutSnakeToCamel())
 	f, err := parser.ParseFile(g.fileSet, "TestRequiredErrors", input, parser.ParseComments)
 	assert.Nil(t, err, "Error parsing no struct input")
 
@@ -38,8 +38,7 @@ func TestNoStructFile(t *testing.T) {
 
 // TestNoFile
 func TestNoFile(t *testing.T) {
-	g := NewGenerator().
-		WithoutSnakeToCamel()
+	g := NewGenerator(WithoutSnakeToCamel())
 	// Parse the file given in arguments
 	_, err := g.GenerateFromFile("")
 	assert.NotNil(t, err, "Error generating formatted code")
@@ -47,12 +46,13 @@ func TestNoFile(t *testing.T) {
 
 // TestExampleFile
 func TestExampleFile(t *testing.T) {
-	g := NewGenerator().
-		WithMarshal().
-		WithSQLDriver().
-		WithCaseInsensitiveParse().
-		WithNames().
-		WithoutSnakeToCamel()
+	g := NewGenerator(
+		WithMarshal(),
+		WithSQLDriver(),
+		WithCaseInsensitiveParse(),
+		WithNames(),
+		WithoutSnakeToCamel(),
+	)
 	// Parse the file given in arguments
 	imported, err := g.GenerateFromFile(testExample)
 	require.Nil(t, err, "Error generating formatted code")
@@ -65,17 +65,76 @@ func TestExampleFile(t *testing.T) {
 	}
 }
 
+// TestExampleFileEmptyHeaders tests the generator with empty header values
+func TestExampleFileEmptyHeaders(t *testing.T) {
+	g := NewGenerator(
+		WithMarshal(),
+		WithSQLDriver(),
+		WithCaseInsensitiveParse(),
+		WithNames(),
+		WithoutSnakeToCamel(),
+	)
+	g.Version = ""
+	g.Revision = ""
+	g.BuildDate = ""
+	g.BuiltBy = ""
+	// Parse the file given in arguments
+	imported, err := g.GenerateFromFile(testExample)
+	require.Nil(t, err, "Error generating formatted code")
+
+	outputLines := strings.Split(string(imported), "\n")
+	cupaloy.SnapshotT(t, outputLines)
+
+	const maxLinesToCheck = 10
+	// Check that the first 10 lines do not contain version information
+	// or build information
+	for idx := range min(maxLinesToCheck, len(outputLines)) {
+		assert.NotContains(t, outputLines[idx], "Version:")
+		assert.NotContains(t, outputLines[idx], "Revision:")
+		assert.NotContains(t, outputLines[idx], "BuildDate:")
+		assert.NotContains(t, outputLines[idx], "BuiltBy:")
+	}
+
+	if false {
+		fmt.Println(string(imported))
+	}
+}
+
 // TestExampleFileMoreOptions
 func TestExampleFileMoreOptions(t *testing.T) {
-	g := NewGenerator().
-		WithMarshal().
-		WithSQLDriver().
-		WithCaseInsensitiveParse().
-		WithNames().
-		WithoutSnakeToCamel().
-		WithMustParse().
-		WithForceLower().
-		WithTemplates(`../example/user_template.tmpl`)
+	g := NewGenerator(
+		WithMarshal(),
+		WithSQLDriver(),
+		WithCaseInsensitiveParse(),
+		WithNames(),
+		WithoutSnakeToCamel(),
+		WithMustParse(),
+		WithForceLower(),
+		WithTemplates(`../example/user_template.tmpl`),
+	)
+	// Parse the file given in arguments
+	imported, err := g.GenerateFromFile(testExample)
+	require.Nil(t, err, "Error generating formatted code")
+
+	outputLines := strings.Split(string(imported), "\n")
+	cupaloy.SnapshotT(t, outputLines)
+
+	if false {
+		fmt.Println(string(imported))
+	}
+}
+
+// TestExampleFileMoreOptionsWithForceUpper — test with force upper option
+func TestExampleFileMoreOptionsWithForceUpper(t *testing.T) {
+	g := NewGenerator(
+		WithMarshal(),
+		WithSQLDriver(),
+		WithNames(),
+		WithoutSnakeToCamel(),
+		WithMustParse(),
+		WithForceUpper(),
+		WithTemplates(`../example/user_template.tmpl`),
+	)
 	// Parse the file given in arguments
 	imported, err := g.GenerateFromFile(testExample)
 	require.Nil(t, err, "Error generating formatted code")
@@ -90,12 +149,13 @@ func TestExampleFileMoreOptions(t *testing.T) {
 
 // TestExampleFile
 func TestNoPrefixExampleFile(t *testing.T) {
-	g := NewGenerator().
-		WithMarshal().
-		WithLowercaseVariant().
-		WithNoPrefix().
-		WithFlag().
-		WithoutSnakeToCamel()
+	g := NewGenerator(
+		WithMarshal(),
+		WithLowercaseVariant(),
+		WithNoPrefix(),
+		WithFlag(),
+		WithoutSnakeToCamel(),
+	)
 	// Parse the file given in arguments
 	imported, err := g.GenerateFromFile(testExample)
 	require.Nil(t, err, "Error generating formatted code")
@@ -108,15 +168,56 @@ func TestNoPrefixExampleFile(t *testing.T) {
 	}
 }
 
+func TestNoIotaExampleFile(t *testing.T) {
+	g := NewGenerator(
+		WithMarshal(),
+		WithLowercaseVariant(),
+		WithNoIota(),
+		WithFlag(),
+		WithoutSnakeToCamel(),
+	)
+	// Parse the file given in arguments
+	imported, err := g.GenerateFromFile(testExample)
+	require.Nil(t, err, "Error generating formatted code")
+
+	outputLines := strings.Split(string(imported), "\n")
+	cupaloy.SnapshotT(t, outputLines)
+
+	if false {
+		fmt.Println(string(imported))
+	}
+}
+
+func TestNoIotaOnlyExampleFile(t *testing.T) {
+	g := NewGenerator(
+		WithMarshal(),
+		WithLowercaseVariant(),
+		WithNoIota(),
+		WithFlag(),
+		WithoutSnakeToCamel(),
+	)
+	// Parse the file given in arguments
+	imported, err := g.GenerateFromFile(testExampleNoIota)
+	require.Nil(t, err, "Error generating formatted code")
+
+	outputLines := strings.Split(string(imported), "\n")
+	cupaloy.SnapshotT(t, outputLines)
+
+	if false {
+		fmt.Println(string(imported))
+	}
+}
+
 // TestExampleFile
 func TestReplacePrefixExampleFile(t *testing.T) {
-	g := NewGenerator().
-		WithMarshal().
-		WithLowercaseVariant().
-		WithNoPrefix().
-		WithPrefix("MyPrefix_").
-		WithFlag().
-		WithoutSnakeToCamel()
+	g := NewGenerator(
+		WithMarshal(),
+		WithLowercaseVariant(),
+		WithNoPrefix(),
+		WithPrefix("MyPrefix_"),
+		WithFlag(),
+		WithoutSnakeToCamel(),
+	)
 	// Parse the file given in arguments
 	imported, err := g.GenerateFromFile(testExample)
 	require.Nil(t, err, "Error generating formatted code")
@@ -131,11 +232,12 @@ func TestReplacePrefixExampleFile(t *testing.T) {
 
 // TestExampleFile
 func TestNoPrefixExampleFileWithSnakeToCamel(t *testing.T) {
-	g := NewGenerator().
-		WithMarshal().
-		WithLowercaseVariant().
-		WithNoPrefix().
-		WithFlag()
+	g := NewGenerator(
+		WithMarshal(),
+		WithLowercaseVariant(),
+		WithNoPrefix(),
+		WithFlag(),
+	)
 
 	// Parse the file given in arguments
 	imported, err := g.GenerateFromFile(testExample)
@@ -151,16 +253,17 @@ func TestNoPrefixExampleFileWithSnakeToCamel(t *testing.T) {
 
 // TestCustomPrefixExampleFile
 func TestCustomPrefixExampleFile(t *testing.T) {
-	g := NewGenerator().
-		WithMarshal().
-		WithLowercaseVariant().
-		WithNoPrefix().
-		WithFlag().
-		WithoutSnakeToCamel().
-		WithPtr().
-		WithSQLNullInt().
-		WithSQLNullStr().
-		WithPrefix("Custom_prefix_")
+	g := NewGenerator(
+		WithMarshal(),
+		WithLowercaseVariant(),
+		WithNoPrefix(),
+		WithFlag(),
+		WithoutSnakeToCamel(),
+		WithPtr(),
+		WithSQLNullInt(),
+		WithSQLNullStr(),
+		WithPrefix("Custom_prefix_"),
+	)
 	// Parse the file given in arguments
 	imported, err := g.GenerateFromFile(testExample)
 	require.Nil(t, err, "Error generating formatted code")
@@ -218,8 +321,7 @@ func TestAliasParsing(t *testing.T) {
 
 	for name, tc := range tests {
 		t.Run(name, func(t *testing.T) {
-			replacementNames = map[string]string{}
-			err := ParseAliases(tc.input)
+			replacementNames, err := ParseAliases(tc.input)
 			if tc.err != nil {
 				require.Error(t, err)
 				require.EqualError(t, err, tc.err.Error())
@@ -244,8 +346,7 @@ func TestEnumParseFailure(t *testing.T) {
 	//}
 	type Animal int
 	`
-	g := NewGenerator().
-		WithoutSnakeToCamel()
+	g := NewGenerator(WithoutSnakeToCamel())
 	f, err := parser.ParseFile(g.fileSet, "TestRequiredErrors", input, parser.ParseComments)
 	assert.Nil(t, err, "Error parsing no struct input")
 
@@ -265,8 +366,7 @@ func TestUintInvalidParsing(t *testing.T) {
 	//)
 	type Animal uint
 	`
-	g := NewGenerator().
-		WithoutSnakeToCamel()
+	g := NewGenerator(WithoutSnakeToCamel())
 	f, err := parser.ParseFile(g.fileSet, "TestRequiredErrors", input, parser.ParseComments)
 	assert.Nil(t, err, "Error parsing no struct input")
 
@@ -286,8 +386,7 @@ func TestIntInvalidParsing(t *testing.T) {
 	//)
 	type Animal int
 	`
-	g := NewGenerator().
-		WithoutSnakeToCamel()
+	g := NewGenerator(WithoutSnakeToCamel())
 	f, err := parser.ParseFile(g.fileSet, "TestRequiredErrors", input, parser.ParseComments)
 	assert.Nil(t, err, "Error parsing no struct input")
 
@@ -305,9 +404,12 @@ func TestAliasing(t *testing.T) {
 	// ENUM(a,b,CDEF) with some extra text
 	type Animal int
 	`
-	g := NewGenerator().
-		WithoutSnakeToCamel()
-	_ = ParseAliases([]string{"CDEF:C"})
+	aliases, err := ParseAliases([]string{"CDEF:C"})
+	require.NoError(t, err)
+	g := NewGenerator(
+		WithoutSnakeToCamel(),
+		WithAliases(aliases),
+	)
 	f, err := parser.ParseFile(g.fileSet, "TestRequiredErrors", input, parser.ParseComments)
 	assert.Nil(t, err, "Error parsing no struct input")
 
@@ -365,4 +467,204 @@ func TestQuotedStrings(t *testing.T) {
 	if false { // Debugging statement
 		fmt.Println(string(output))
 	}
+}
+
+func TestStringWithSingleDoubleQuoteValue(t *testing.T) {
+	input := `package test
+	// ENUM(DoubleQuote='"')
+	type Char string
+	`
+	g := NewGenerator()
+	f, err := parser.ParseFile(g.fileSet, "TestRequiredErrors", input, parser.ParseComments)
+	assert.Nil(t, err, "Error parsing no struct input")
+
+	output, err := g.Generate(f)
+	assert.Nil(t, err, "Error generating formatted code")
+	assert.Contains(t, string(output), "CharDoubleQuote Char = \"\\\"\"")
+	if false { // Debugging statement
+		fmt.Println(string(output))
+	}
+}
+
+func TestStringWithSingleSingleQuoteValue(t *testing.T) {
+	input := `package test
+	// ENUM(SingleQuote="'")
+	type Char string
+	`
+	g := NewGenerator()
+	f, err := parser.ParseFile(g.fileSet, "TestRequiredErrors", input, parser.ParseComments)
+	assert.Nil(t, err, "Error parsing no struct input")
+
+	output, err := g.Generate(f)
+	assert.Nil(t, err, "Error generating formatted code")
+	assert.Contains(t, string(output), "CharSingleQuote Char = \"'\"")
+	if false { // Debugging statement
+		fmt.Println(string(output))
+	}
+}
+
+func TestStringWithSingleBacktickValue(t *testing.T) {
+	input := `package test
+	// ENUM(SingleQuote="` + "`" + `")
+	type Char string
+	`
+	g := NewGenerator()
+	f, err := parser.ParseFile(g.fileSet, "TestRequiredErrors", input, parser.ParseComments)
+	assert.Nil(t, err, "Error parsing no struct input")
+
+	output, err := g.Generate(f)
+	assert.Nil(t, err, "Error generating formatted code")
+	assert.Contains(t, string(output), "CharSingleQuote Char = \"`\"")
+	if false { // Debugging statement
+		fmt.Println(string(output))
+	}
+}
+
+// TestNewGeneratorWithConfig tests the NewGeneratorWithConfig constructor
+func TestNewGeneratorWithConfig(t *testing.T) {
+	config := GeneratorConfig{
+		NoPrefix:        true,
+		LowercaseLookup: true,
+		Marshal:         true,
+		SQL:             true,
+		SQLInt:          true,
+		JSONPkg:         "custom/json",
+		Prefix:          "TestPrefix",
+		BuildTags:       []string{"tag1", "tag2"},
+		NoComments:      true,
+		Values:          true,
+	}
+
+	g := NewGeneratorWithConfig(config)
+
+	assert.NotNil(t, g)
+	assert.Equal(t, config.NoPrefix, g.NoPrefix)
+	assert.Equal(t, config.LowercaseLookup, g.LowercaseLookup)
+	assert.Equal(t, config.Marshal, g.Marshal)
+	assert.Equal(t, config.SQL, g.SQL)
+	assert.Equal(t, config.SQLInt, g.SQLInt)
+	assert.Equal(t, config.JSONPkg, g.JSONPkg)
+	assert.Equal(t, config.Prefix, g.Prefix)
+	assert.Equal(t, config.BuildTags, g.BuildTags)
+	assert.Equal(t, config.NoComments, g.NoComments)
+	assert.Equal(t, config.Values, g.Values)
+
+	// Test default values
+	assert.Equal(t, "-", g.Version)
+	assert.Equal(t, "-", g.Revision)
+	assert.Equal(t, "-", g.BuildDate)
+	assert.Equal(t, "-", g.BuiltBy)
+	assert.NotNil(t, g.knownTemplates)
+	assert.NotNil(t, g.t)
+	assert.NotNil(t, g.fileSet)
+	assert.NotNil(t, g.userTemplateNames)
+}
+
+// TestNewGeneratorConfig tests the NewGeneratorConfig constructor
+func TestNewGeneratorConfig(t *testing.T) {
+	config := NewGeneratorConfig()
+
+	assert.NotNil(t, config)
+	assert.False(t, config.NoPrefix)
+	assert.NotNil(t, config.ReplacementNames)
+	assert.Equal(t, "encoding/json", config.JSONPkg)
+	assert.Empty(t, config.ReplacementNames)
+}
+
+// TestWithSQLInt tests the WithSQLInt option
+func TestWithSQLInt(t *testing.T) {
+	config := &GeneratorConfig{}
+	option := WithSQLInt()
+
+	assert.False(t, config.SQLInt)
+	option(config)
+	assert.True(t, config.SQLInt)
+}
+
+// TestWithValues tests the WithValues option
+func TestWithValues(t *testing.T) {
+	config := &GeneratorConfig{}
+	option := WithValues()
+
+	assert.False(t, config.Values)
+	option(config)
+	assert.True(t, config.Values)
+}
+
+// TestWithJsonPkg tests the WithJsonPkg option
+func TestWithJsonPkg(t *testing.T) {
+	config := &GeneratorConfig{}
+	testPkg := "custom/json/package"
+	option := WithJsonPkg(testPkg)
+
+	assert.Empty(t, config.JSONPkg)
+	option(config)
+	assert.Equal(t, testPkg, config.JSONPkg)
+}
+
+// TestWithNoComments tests the WithNoComments option
+func TestWithNoComments(t *testing.T) {
+	config := &GeneratorConfig{}
+	option := WithNoComments()
+
+	assert.False(t, config.NoComments)
+	option(config)
+	assert.True(t, config.NoComments)
+}
+
+// TestWithBuildTags tests the WithBuildTags option
+func TestWithBuildTags(t *testing.T) {
+	config := &GeneratorConfig{}
+	testTags := []string{"tag1", "tag2", "tag3"}
+	option := WithBuildTags(testTags...)
+
+	assert.Empty(t, config.BuildTags)
+	option(config)
+	assert.Equal(t, testTags, config.BuildTags)
+
+	// Test appending more tags
+	moreTags := []string{"tag4", "tag5"}
+	option2 := WithBuildTags(moreTags...)
+	option2(config)
+	expectedTags := append(testTags, moreTags...)
+	assert.Equal(t, expectedTags, config.BuildTags)
+}
+
+// TestAllOptionsIntegration tests using multiple options together
+func TestAllOptionsIntegration(t *testing.T) {
+	g := NewGenerator(
+		WithSQLInt(),
+		WithValues(),
+		WithJsonPkg("custom/json"),
+		WithNoComments(),
+		WithBuildTags("integration", "test"),
+	)
+
+	assert.True(t, g.SQLInt)
+	assert.True(t, g.Values)
+	assert.Equal(t, "custom/json", g.JSONPkg)
+	assert.True(t, g.NoComments)
+	assert.Equal(t, []string{"integration", "test"}, g.BuildTags)
+}
+
+// TestGeneratorConfigWithTemplates tests NewGeneratorWithConfig with templates
+func TestGeneratorConfigWithTemplates(t *testing.T) {
+	config := GeneratorConfig{
+		// Use empty template file names to avoid file not found errors
+		TemplateFileNames: []string{},
+	}
+
+	g := NewGeneratorWithConfig(config)
+	assert.NotNil(t, g)
+	assert.Equal(t, config.TemplateFileNames, g.TemplateFileNames)
+
+	// Test with non-empty but valid scenario - no actual templates needed for coverage
+	config2 := GeneratorConfig{
+		NoPrefix: true,
+		Values:   true,
+	}
+	g2 := NewGeneratorWithConfig(config2)
+	assert.NotNil(t, g2)
+	assert.True(t, g2.NoPrefix)
+	assert.True(t, g2.Values)
 }
